Create Mission.js after refactor code
